Megalogic/requirements.txt
streamlit==1.35.0
mysql-connector==2.2.9
pandas==2.0.3
mlxtend==0.19.0
seaborn==0.11.2
networkx==2.5

